+++ src/main/kotlin/com/chattriggers/ctjs/launch/plugin/guiScreen.kt
@@ -18,9 +18,6 @@ import org.lwjgl.input.Keyboard
 fun injectGuiScreen() {
     injectSendChatMessage()
     injectHandleKeyboardInput()
-    injectMouseClick()
-    injectMouseRelease()
-    injectMouseDrag()
     injectTextComponentClick()
     injectTextComponentHover()
     injectRenderTooltip()
@@ -80,111 +77,6 @@ fun injectHandleKeyboardInput() = inject {
                 event
             )
 
-            if (event.isCancelled())
-                methodReturn()
-        }
-    }
-}
-
-fun injectMouseClick() = inject {
-    className = "net/minecraft/client/gui/GuiScreen"
-    methodName = "handleMouseInput"
-    methodDesc = "()V"
-
-    at = At(
-        InjectionPoint.INVOKE(
-            Descriptor(
-                "net/minecraft/client/gui/GuiScreen",
-                "mouseClicked",
-                "(III)V"
-            )
-        )
-    )
-
-    methodMaps = mapOf(
-        "func_146274_d" to "handleMouseInput",
-        "func_73864_a" to "mouseClicked"
-    )
-
-    codeBlock {
-        val local0 = shadowLocal<GuiScreen>()
-        val local1 = shadowLocal<Int>()
-        val local2 = shadowLocal<Int>()
-        val local3 = shadowLocal<Int>()
-
-        code {
-            val event = CancellableEvent()
-            TriggerType.GuiMouseClick.triggerAll(local1, local2, local3, local0, event)
-            if (event.isCancelled())
-                methodReturn()
-        }
-    }
-}
-
-fun injectMouseRelease() = inject {
-    className = "net/minecraft/client/gui/GuiScreen"
-    methodName = "handleMouseInput"
-    methodDesc = "()V"
-
-    at = At(
-        InjectionPoint.INVOKE(
-            Descriptor(
-                "net/minecraft/client/gui/GuiScreen",
-                "mouseReleased",
-                "(III)V"
-            )
-        )
-    )
-
-    methodMaps = mapOf(
-        "func_146274_d" to "handleMouseInput",
-        "func_146286_b" to "mouseReleased"
-    )
-
-    codeBlock {
-        val local0 = shadowLocal<GuiScreen>()
-        val local1 = shadowLocal<Int>()
-        val local2 = shadowLocal<Int>()
-        val local3 = shadowLocal<Int>()
-
-        code {
-            val event = CancellableEvent()
-            TriggerType.GuiMouseRelease.triggerAll(local1, local2, local3, local0, event)
-            if (event.isCancelled())
-                methodReturn()
-        }
-    }
-}
-
-fun injectMouseDrag() = inject {
-    className = "net/minecraft/client/gui/GuiScreen"
-    methodName = "handleMouseInput"
-    methodDesc = "()V"
-
-    at = At(
-        InjectionPoint.INVOKE(
-            Descriptor(
-                "net/minecraft/client/gui/GuiScreen",
-                "mouseClickMove",
-                "(IIIJ)V"
-            )
-        )
-    )
-
-    methodMaps = mapOf(
-        "func_146274_d" to "handleMouseInput",
-        "func_146273_a" to "mouseClickMove"
-    )
-
-    codeBlock {
-        val local0 = shadowLocal<GuiScreen>()
-        val local1 = shadowLocal<Int>()
-        val local2 = shadowLocal<Int>()
-        val local3 = shadowLocal<Int>()
-
-        code {
-            val event = CancellableEvent()
-            TriggerType.GuiMouseDrag.triggerAll(local1, local2, local3, local0, event)
             if (event.isCancelled())
                 methodReturn()
         }